override max users
